requirements.txt
beautifulsoup4==4.4.1
lxml==3.6.4
py==1.4.31
pytest==2.9.2
python-dateutil==2.6.1
pytz==2017.2
requests==2.10.0
requests-mock==1.3.0
six==1.10.0
tzlocal==1.4
